requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
statsmodels>=0.14.0
arch>=6.3.0
yfinance>=0.2.18
python-bcb>=0.2.0
matplotlib>=3.7.1
seaborn>=0.12.2
cvxpy>=1.3.1
scikit-learn>=1.2.2
plotly>=5.14.1
ipython
jupyter
